requirements.txt
bcrypt==4.2.1
flask==3.1.2
psycopg2-binary==2.9.11
PyJWT==2.10.1
boto3